Random seed in resource
Extending resource with some random characters+digits may be helpful in case when only a single account (JID) is used for connecting to a conference room. In my case, Prosody.im kicked the old user leaving alone the new user in the room. It's also possible to change it in the server configuration but the convention is that each connection should have its unique resource.