Fix #204 - add support for XHTML
Together with colors-xhtml in the candy-plugins repository (branch
feature/xhtml-colors) this will enable Candy to show validated
XHTML if enabled with the `enableXHTML` config flag.

In order to make this work, I had to fix the muc plugin of strophe.
Until strophe/strophejs-plugins#14 is not merged, this is done manually.