Remove generated assets
These files are generated from true source contained in this repo and are only relevant to a *release* of Candy, not to development. Leaving these out of source control is semantically correct (they are not source, they are release artefacts) and will make contribution easier because commits will cleanly apply out of order more often (changes to generated source almost never cleanly compose from the same start-point due to minification).