Add timestamp data to messages
This will allow JS code to look at messages already rendered in the DOM and be able to tell the original time the message was received. Among other things, this can be used to determine where to start loading more history, based on which messages are already in the scroll back.