Add domain list support.
First pass at adding support for a domain selector on the login form.

Adds a new core parameter 'domains' which should be a list of
acceptable domains.  Only the first whitespace-delineated 'word' will
be used, so additional information can be added to the domains (eg
"chat.foo.com (for general users").

When defined, the login logic will strip off any user-provided domain
and append the selected domain to the JID before attempting to log
into the server.

Problems:
        - Doubles the width of all modal dialogue boxes, not just the
        - Candy.View.Pane.showLoginForm() pulls the domain list
          directly via Candy.Core.getOptions() rather than taking
          them as a parameter.  This is probably not 'best practice'.